Renamed `fetch.py` -> `utils.py` and removed unused functions.
Additionally, all exceptions were moved into `exceptions.py` and
some errors were corrected in `summarizer.py`.
